Renamed routing files to follow kabab case
